pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "embedded-trainer"
version = "0.1.0"
description = "A gamified terminal app for learning embedded systems fundamentals"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.80.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
embedded-trainer = "embedded_trainer.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
